requirements.txt
ago==0.0.95
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
anyio==4.0.0
async-timeout==4.0.3
attrs==22.1.0
Automat==22.10.0
bardapi==0.1.33
beautifulsoup4==4.12.2
boto3==1.28.48
botocore==1.31.48
browser-cookie3==0.19.1
bs4==0.0.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.3
cssselect==1.2.0
decorator==4.4.2
deep-translator==1.11.4
dotmap==1.3.30
elastic-transport==8.4.0
elasticsearch==8.9.0
faust-cchardet==2.1.19
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.4
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.1
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
googleapis-common-protos==1.60.0
greenlet==2.0.1
grpcio==1.58.0
grpcio-status==1.58.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hjson==3.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hurry.filesize==0.9
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.9
importlib-metadata==5.1.0
importlib-resources==5.10.1
incremental==22.10.0
iniconfig==1.1.1
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.17.3
langdetect==1.0.9
linkedin-api-client==0.3.0
lxml==4.9.3
lz4==4.3.2
MarkupSafe==2.1.1
midjourney-py==1.0.4
moviepy==1.0.3
multidict==6.0.4
news-please==1.5.35
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.25.2
oauthlib==3.2.2
openai==0.28.1
packaging==22.0
parsel==1.8.1
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
plac==1.3.5
pluggy==1.0.0
proglog==0.1.10
Protego==0.3.0
proto-plus==1.22.3
protobuf==4.24.3
psycopg2-binary==2.9.7
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.18.0
PyDispatcher==2.0.7
PyJWT==2.6.0
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pyrsistent==0.19.2
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.6
PyYAML==6.0.1
queuelib==1.6.2
readability-lxml==0.8.1
regex==2023.8.8
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.2
Scrapy==2.10.1
service-identity==23.1.0
sgmllib3k==1.0.0
six==1.16.0
snakecase==1.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.46
tinysegmenter==0.3
tldextract==3.5.0
tomli==2.0.1
tqdm==4.66.1
tweepy==4.14.0
Twisted==22.10.0
typing_extensions==4.7.1
urllib3==1.26.13
w3lib==2.1.2
warcio==1.7.4
Werkzeug==2.1.2
yarl==1.9.2
zipp==3.11.0
zope.interface==6.0
